added deleteById to performanceManagement

+++ src/main/java/com/example/cpm/Service/PerformanceService.java
@@ -1,6 +1,7 @@
 package com.example.cpm.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -23,19 +24,39 @@ public class PerformanceService {
 
     public String addPerformance(Performance performance) {
         Talent existingTalent = talentRepository.findById(performance.getTalentId()).orElseThrow(() -> new IllegalArgumentException("Talent not found"));
+        
+        Optional<Performance> existingPerformance = performanceRepo.findById(performance.getTalentId());
+        if (existingPerformance.isPresent()) 
+            return "Performance with this ID already exists";
 
         performance.setTalentName(existingTalent.getTalentName());
         performance.setEkYear(existingTalent.getEkYear());
         performance.setTalentSkills(existingTalent.getTalentSkills());
         performanceRepo.save(performance);
-        return "saved";
+        return "Performance saved successfully";
     }
 
     public void updateFeedback(Performance performance) {
+        Optional<Performance> existingPerformance = performanceRepo.findById(performance.getTalentId());
+        if (existingPerformance.isEmpty()) 
+            throw new IllegalArgumentException("Performance with ID " + performance.getTalentId() + " does not exist.");
+
         performanceRepo.save(performance);
     }
 
     public void deletePerformance(Performance performance) {
+        Optional<Performance> existingPerformance = performanceRepo.findById(performance.getTalentId());
+        if (existingPerformance.isEmpty()) 
+            throw new IllegalArgumentException("Performance with ID " + performance.getTalentId() + " does not exist.");
+
         performanceRepo.delete(performance);
+    }
+
+    public void deletePerformance(int talentId){
+        Optional<Performance> existingPerformance = performanceRepo.findById(talentId);
+        if (existingPerformance.isEmpty()) 
+            throw new IllegalArgumentException("Performance with ID " + talentId + " does not exist.");
+        
+        performanceRepo.deleteById(talentId);
     }
 }
